QPID-8615: [Broker-J] JUnit 5 migration (#165)

+++ perftests/src/test/java/org/apache/qpid/disttest/controller/config/SessionConfigTest.java
@@ -21,6 +21,8 @@ package org.apache.qpid.disttest.controller.config;
 
 import static org.apache.qpid.disttest.controller.config.ConfigTestUtils.assertCommandEquals;
 import static org.apache.qpid.disttest.controller.config.ConfigTestUtils.getCommand;
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertNotNull;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;
 
@@ -29,29 +31,14 @@ import java.util.List;
 
 import javax.jms.Session;
 
-import org.junit.Assert;
-
 import org.apache.qpid.disttest.message.Command;
 import org.apache.qpid.disttest.message.CreateConsumerCommand;
 import org.apache.qpid.disttest.message.CreateProducerCommand;
 import org.apache.qpid.disttest.message.CreateSessionCommand;
 
-import org.junit.Assert;
-import org.junit.Before;
-import org.junit.After;
-import org.junit.Test;
+import org.junit.jupiter.api.Test;
 
 import org.apache.qpid.test.utils.UnitTestBase;
-import static org.junit.Assert.assertSame;
-import static org.junit.Assert.assertNotSame;
-import static org.junit.Assert.fail;
-import static org.junit.Assert.assertNotNull;
-
-import static org.junit.Assert.assertFalse;
-import static org.junit.Assert.assertTrue;
-
-import static org.junit.Assert.assertEquals;
-import static org.junit.Assert.assertNull;
 
 public class SessionConfigTest extends UnitTestBase
 {
@@ -71,14 +58,14 @@ public class SessionConfigTest extends UnitTestBase
         SessionConfig sessionConfig = createSessionConfigWithChildCommands();
 
         List<Command> commands = sessionConfig.createCommands(CONNECTION_NAME);
-        assertEquals((long) 3, (long) commands.size());
+        assertEquals(3, commands.size());
 
         assertCommandEquals(commands, 0, CreateSessionCommand.class);
         assertCommandEquals(commands, 1, CreateProducerCommand.class);
         assertCommandEquals(commands, 2, CreateConsumerCommand.class);
 
         CreateSessionCommand createSessionCommand = getCommand(commands, 0);
-        assertEquals((long) Session.AUTO_ACKNOWLEDGE, (long) createSessionCommand.getAcknowledgeMode());
+        assertEquals(Session.AUTO_ACKNOWLEDGE, createSessionCommand.getAcknowledgeMode());
         assertEquals(SESSION, createSessionCommand.getSessionName());
         assertEquals(CONNECTION_NAME, createSessionCommand.getConnectionName());
     }
@@ -87,7 +74,7 @@ public class SessionConfigTest extends UnitTestBase
     public void testGetTotalNumberOfParticipants()
     {
         SessionConfig sessionConfig = createSessionConfigWithOneConsumerAndOneProducer();
-        assertEquals((long) 2, (long) sessionConfig.getTotalNumberOfParticipants());
+        assertEquals(2, sessionConfig.getTotalNumberOfParticipants());
     }
 
     private SessionConfig createSessionConfigWithOneConsumerAndOneProducer()
@@ -108,6 +95,4 @@ public class SessionConfigTest extends UnitTestBase
                                  Collections.singletonList(consumerConfig),
                                  Collections.singletonList(producerConfig));
     }
-
-
 }
